ref(android): rename the Android app
