+++ in-room-controller/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.mobile",
+    package = "org.jitsi.spot",
 )
 
 android_resource(
     name = "res",
-    package = "com.mobile",
+    package = "org.jitsi.spot",
     res = "src/main/res",
 )
