+++ in-room-controller/android/app/src/main/AndroidManifest.xml
@@ -1,11 +1,11 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.mobile">
+    package="org.jitsi.spot">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW"/>
 
     <application
-      android:name=".MainApplication"
+      android:name="org.jitsi.spot.MainApplication"
       android:label="@string/app_name"
       android:icon="@mipmap/ic_launcher"
       android:roundIcon="@mipmap/ic_launcher_round"
@@ -13,7 +13,7 @@
       android:theme="@style/AppTheme"
       android:usesCleartextTraffic="true">
       <activity
-        android:name=".MainActivity"
+        android:name="org.jitsi.spot.MainActivity"
         android:label="@string/app_name"
         android:configChanges="keyboard|keyboardHidden|orientation|screenSize"
         android:windowSoftInputMode="adjustResize">
