+++ in-room-controller/android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'mobile'
+rootProject.name = 'in-room-controller'
 include ':@react-native-community_async-storage'
 project(':@react-native-community_async-storage').projectDir = new File(rootProject.projectDir, '../node_modules/@react-native-community/async-storage/android')
 include ':react-native-webview'
